+++ src/bridges/ethereum-darwinia/model/index.ts
@@ -1,0 +1,1 @@
+export {};
