+++ src/bridges/ethereum-darwiniaDVM/DarwiniaDVM2EthereumRecord.tsx
@@ -1,0 +1,3 @@
+export function DarwiniaDVM2EthereumRecord() {
+  return <span>DarwiniaDVM2EthereumRecord</span>;
+}
