+++ src/bridges/ethereum-darwiniaDVM/Ethereum2DarwiniaDVMRecord.tsx
@@ -1,0 +1,3 @@
+export function Ethereum2DarwiniaDVMRecord() {
+  return <span>Ethereum2DarwiniaDVMRecord</span>;
+}
