+++ src/bridges/ethereum-darwiniaDVM/index.ts
@@ -1,0 +1,4 @@
+export * from './Ethereum2DarwiniaDVM';
+export * from './DarwiniaDVM2Ethereum';
+export * from './Ethereum2DarwiniaDVMRecord';
+export * from './DarwiniaDVM2EthereumRecord';
