--- src/bridges/ethereum-dvm/Dvm2Ethereum.tsx
@@ -1,3 +1,0 @@
-export function Dvm2Ethereum() {
-  return <span>Dvm2Ethereum</span>;
-}
