--- src/bridges/ethereum-dvm/Dvm2EthereumRecord.tsx
@@ -1,3 +1,0 @@
-export function Dvm2EthereumRecord() {
-  return <span>Dvm2EthereumRecord</span>;
-}
