--- src/bridges/ethereum-dvm/Ethereum2Dvm.tsx
@@ -1,3 +1,0 @@
-export function Ethereum2Dvm() {
-  return <span>Ethereum2Dvm</span>;
-}
