--- src/bridges/ethereum-dvm/Ethereum2DvmRecord.tsx
@@ -1,3 +1,0 @@
-export function Ethereum2DvmRecord() {
-  return <span>Ethereum2DvmRecord</span>;
-}
