--- src/bridges/ethereum-dvm/index.ts
@@ -1,4 +1,0 @@
-export * from './Ethereum2Dvm';
-export * from './Dvm2Ethereum';
-export * from './Ethereum2DvmRecord';
-export * from './Dvm2EthereumRecord';
