+++ src/bridges/substrate-dvm/config/index.ts
@@ -1,0 +1,1 @@
+export {};
