+++ src/bridges/substrate-dvm/hooks/index.ts
@@ -1,0 +1,1 @@
+export {};
