+++ src/bridges/substrate-dvm/index.ts
@@ -1,4 +1,2 @@
-export * from './Substrate2Dvm';
-export * from './Dvm2Substrate';
-export * from './Substrate2DvmRecord';
-export * from './Dvm2SubstrateRecord';
+export * from './DVM2Substrate';
+export * from './Substrate2DVM';
