+++ src/bridges/substrate-substrateDVM/config/index.ts
@@ -1,0 +1,1 @@
+export {};
