+++ src/bridges/substrate-substrateDVM/hooks/index.ts
@@ -1,0 +1,1 @@
+export {};
