+++ src/bridges/substrate-substrateDVM/index.ts
@@ -1,4 +1,2 @@
 export * from './Substrate2SubstrateDVM';
 export * from './SubstrateDVM2Substrate';
-export * from './Substrate2SubstrateDVMRecord';
-export * from './SubstrateDVM2SubstrateRecord';
