--- src/components/bridge/ethereum-dvm/index.ts
@@ -1,2 +1,0 @@
-export * from './DarwiniaDVM2Ethereum';
-export * from './Ethereum2DarwiniaDVM';
