--- src/components/bridge/substrate-dvm/index.ts
@@ -1,2 +1,0 @@
-export * from './DVM2Substrate';
-export * from './Substrate2DVM';
