--- src/components/bridge/substrate-substrateDVM/index.ts
@@ -1,2 +1,0 @@
-export * from './Substrate2SubstrateDVM';
-export * from './SubstrateDVM2Substrate';
